quits all jobs on exit, waits for childs and updates info

+++ Pysrcs/tools.py
@@ -8,8 +8,9 @@ import errors
 
 def kill_jobs(programList):
 	for program in programList:
-		for pid in program.pidList:
-			os.kill(pid[0], signal.SIGKILL)
+		if program.state != "Finished" and program.state != "Not started":
+			for pid in program.pidList:
+				os.kill(pid[0], signal.SIGKILL)
 
 def verify_config(mode, configList):
 	"""Verifies that the parsed yaml file doesn't contain any errors and has
